+++ thread_hardware.h
@@ -33,7 +33,7 @@ private:
     QTimer *beep_timer;//鸣叫鸣叫定时控制
 
     bool start_heat_high_speed;//该标志表示开启全速加热，并开始进行温度数据记录，全速加热将根据信号中的stop_temp自动停止
-    double stop_temp;
+    int stop_temp;
 
     bool flag_record_temp;
     FILE *fp_record_temp;
@@ -77,7 +77,7 @@ public slots:
     void recei_fro_logic_magnetic(MAGNETIC magnetic_para);
 
     /* 接收到来自逻辑线程的信号后驱使逻辑线程开启全速加热，并开始进行温度数据记录，全速加热将根据信号中的stop_temp自动停止 */
-    void recei_fro_logic_start_heat_high_speed(double stop_temp_para);
+    void recei_fro_logic_start_heat_high_speed(int stop_temp_para);
 
     /* 开始记录温度数据 */
     void recei_fro_logic_start_record_temp();
